requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.0.5
djangorestframework==3.11.1
djongo==1.3.3
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
